Add slaproxy command.
> slaproxy -token bot-token

This will proxy methods marked with "proxy" in endpoints.json to
the actual slack API server. For everything else, we intercept it
and return some mock response.

This allows you to point your app to this server, and retrieve
information from Slack while avoiding to post data to the live
Slack teams